+++ GIP/gip/GeoImage.h
@@ -27,7 +27,7 @@
 #include <gip/GeoRaster.h>
 #include <gip/GeoFeature.h>
 #include <stdint.h>
-#include <gip/Utils.h>
+#include <gip/utils.h>
 
 namespace gip {
     //using std::string;
