+++ GIP/gip/utils.h
